+++ src/contenido/ListarContrato.js
@@ -1,8 +1,9 @@
-import React, { useEffect, useState } from "react";
+import * as React from 'react';
 import axios from "axios";
-import { Link, useParams } from "react-router-dom";
+import { Link } from "react-router-dom";
 import Header from '../Header';
 import NavBar from '../NavBar';
+import Swal from 'sweetalert2';
 
 export default function ListarContrato () {
   
@@ -17,7 +18,24 @@ export default function ListarContrato () {
   },[])
 
   const deleteContrato = async (id) => {
-    await axios.delete(`http://localhost:8080/editc/${id}`);
+    Swal.fire({
+      title: 'Estás seguro de eliminar estos datos?',
+      text: "No podras revertir esto!",
+      icon: 'Advertencia',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Si, eliminar!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+     axios.delete(`http://localhost:8080/editc/${id}`);
+     Swal.fire(
+      'Eliminados!',
+      'Tus datos han sido eliminados.',
+      'Completados'
+    )
+  }
+})
   };
     return (
         
@@ -51,7 +69,7 @@ export default function ListarContrato () {
   {/* Content */}
   <div className="container-fluid">
     <div className="table-responsive">
-      <table className="table table-dark table-sm">
+      <table className="table table-secondary table-sm">
         <thead>
           <tr className="text-center roboto-medium">
             <th>#</th>
